backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
paramiko==3.5.0
pyyaml==6.0.2
python-frontmatter==1.1.0
anthropic==0.42.0
python-multipart==0.0.20
mcp>=1.2.0
